refactor: fixed project details endpoint packaging
