+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/MicroserviceRoutes.scala
@@ -57,7 +57,7 @@ private class MicroserviceRoutes[F[_]: Async](
     queryEndpoint:           QueryEndpoint[F],
     lineageEndpoint:         lineage.Endpoint[F],
     ontologyEndpoint:        ontology.Endpoint[F],
-    projectEndpoint:         projectdetails.Endpoint[F],
+    projectDetailsEndpoint:  projects.details.Endpoint[F],
     projectDatasetsEndpoint: ProjectDatasetsEndpoint[F],
     docsEndpoint:            docs.Endpoint[F],
     usersProjectsEndpoint:   users.projects.Endpoint[F],
@@ -75,7 +75,7 @@ private class MicroserviceRoutes[F[_]: Async](
   import ontologyEndpoint._
   import org.http4s.HttpRoutes
   import projectDatasetsEndpoint._
-  import projectEndpoint._
+  import projectDetailsEndpoint._
   import projectPathAuthorizer.{authorize => authorizePath}
   import queryEndpoint._
   import routesMetrics._
@@ -283,7 +283,7 @@ private object MicroserviceRoutes {
         queryEndpoint           <- QueryEndpoint()
         lineageEndpoint         <- lineage.Endpoint[IO]
         ontologyEndpoint        <- ontology.Endpoint[IO]
-        projectEndpoint         <- projectdetails.Endpoint[IO]
+        projectDetailsEndpoint  <- projects.details.Endpoint[IO]
         projectDatasetsEndpoint <- ProjectDatasetsEndpoint[IO]
         docsEndpoint            <- docs.Endpoint[IO]
         usersProjectsEndpoint   <- users.projects.Endpoint[IO]
@@ -298,7 +298,7 @@ private object MicroserviceRoutes {
                                      queryEndpoint,
                                      lineageEndpoint,
                                      ontologyEndpoint,
-                                     projectEndpoint,
+                                     projectDetailsEndpoint,
                                      projectDatasetsEndpoint,
                                      docsEndpoint,
                                      usersProjectsEndpoint,
