+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/projects/details/EndpointDocs.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.projectdetails
+package io.renku.knowledgegraph.projects.details
 
 import cats.MonadThrow
 import cats.syntax.all._
