+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/projects/details/GitLabProjectFinder.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.projectdetails
+package io.renku.knowledgegraph.projects.details
 
 import GitLabProjectFinder.GitLabProject
 import cats.effect.kernel.Async
