+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/projects/details/ProjectJsonLDEncoder.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.projectdetails
+package io.renku.knowledgegraph.projects.details
 
 import io.renku.graph.model.Schemas._
 import io.renku.graph.model.entities
