+++ knowledge-graph/src/main/scala/io/renku/knowledgegraph/users/projects/model.scala
@@ -26,7 +26,7 @@ import io.renku.config.renku
 import io.renku.graph.model.{RenkuUrl, persons, projects}
 import io.renku.http.rest.Links.{Href, Link, Method, Rel, _links}
 import io.renku.json.JsonOps._
-import io.renku.knowledgegraph.projectdetails
+import io.renku.knowledgegraph.projects.details.Endpoint
 import io.renku.tinytypes.json.TinyTypeEncoders._
 
 private object model {
@@ -67,7 +67,7 @@ private object model {
             .addIfDefined("description" -> project.maybeDesc)
             .deepMerge(
               _links(
-                Link(Rel("details") -> projectdetails.Endpoint.href(renkuApiUrl, project.path))
+                Link(Rel("details") -> Endpoint.href(renkuApiUrl, project.path))
               )
             )
         }
