+++ knowledge-graph/src/test/scala/io/renku/knowledgegraph/projects/details/ProjectFinderSpec.scala
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package io.renku.knowledgegraph.projectdetails
+package io.renku.knowledgegraph.projects.details
 
 import Converters._
 import GitLabProjectFinder.GitLabProject
